Move hours util to app/utils alongside the other date helpers
