+++ app/utils/hours.ts
